+++ src/parsers/problem/LibreOJProblemParser.ts
@@ -5,16 +5,14 @@ import { Parser } from '../Parser';
 
 export class LibreOJProblemParser extends Parser {
   public getMatchPatterns(): string[] {
-    return ['https://loj.ac/problem/*'];
+    return ['https://loj.ac/problem/*', 'https://loj.ac/contest/*/problem/*'];
   }
 
   public async parse(url: string, html: string): Promise<Sendable> {
     const elem = htmlToElement(html);
     const task = new TaskBuilder().setUrl(url);
 
-    const title = elem.querySelector('.ui.header').textContent.trim();
-    task.setName(title.split('. ')[1]);
-
+    task.setName(elem.querySelector('.ui.header').textContent.trim());
     task.setGroup('LibreOJ');
 
     const timeLimitStr = elem.querySelector('.row > .ui.label:nth-child(2)').textContent;
@@ -28,12 +26,14 @@ export class LibreOJProblemParser extends Parser {
       return titleElem !== null && titleElem.textContent === '样例';
     });
 
-    const sampleBlocks = samplesRow.querySelectorAll('pre > code');
-    for (let i = 0; i < sampleBlocks.length; i += 2) {
-      const input = sampleBlocks[i].textContent.trim();
-      const output = sampleBlocks[i + 1].textContent.trim();
+    if (samplesRow) {
+      const sampleBlocks = samplesRow.querySelectorAll('pre > code');
+      for (let i = 0; i < sampleBlocks.length; i += 2) {
+        const input = sampleBlocks[i].textContent.trim();
+        const output = sampleBlocks[i + 1].textContent.trim();
 
-      task.addTest(input, output);
+        task.addTest(input, output);
+      }
     }
 
     return task.build();